[clang-offload-bundler] add -input/-output options
Currently, clang-offload-bundler has -inputs and -outputs options that accept values with comma as the delimiter. This causes issues with file paths containing commas, which are valid file paths on Linux.

This add two new options -input and -output, which accept one single file, and allow multiple instances. This allows arbitrary file paths. The old -inputs and -outputs options will be kept for backward compatibility, but are not allowed to be used with -input and -output options for simplicity. In the future, -inputs and -outputs options will be phasing out.

RFC: https://discourse.llvm.org/t/rfc-adding-input-and-output-options-to-clang-offload-bundler/60049

Differential Revision: https://reviews.llvm.org/D120662

Change-Id: I3cb5f58a025cc9c1a40022e5651d48685a058709